+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Fix settings dialog not working
 - Fix tray show and hide problem
 - Fix tray setting can open multiple dialog
+- Add Free disk space to settings
+- Add Available disk space to settings
+- Add Total disk space to settings
 
 1.0.5
 - Add debounce control